requirements.txt
streamlit
python-dotenv
pypdf
resend
requests
openai
boto3
